corrected structure of the redux-ducks
