+++ src/redux/ducks/order.ts
@@ -1,46 +1,28 @@
 import {
-  Coordinates,
   FetchProductListAction,
-  GetUserCoordinatesAction,
-  GetUserLocationAction,
   OrderActions,
   OrderState,
   ProductList,
   SetProductListAction,
-  UserLocation,
-} from '../types/orderTypes';
+} from '../../types/orderTypes';
+import { call, put, takeEvery } from 'redux-saga/effects';
+import { fetchProducts } from '../../api';
 
 export enum OrderActionTypes {
   SET_PRODUCT_LIST = 'SET_PRODUCT_LIST',
   FETCH_PRODUCT_LIST = 'FETCH_PRODUCT_LIST',
-  GET_USER_LOCATION = 'GET_USER_LOCATION',
-  GET_USER_COORDINATES = 'GET_USER_COORDINATES',
 }
 
 let initialState: OrderState = {
   productList: [],
   shipping: 0,
   taxes: 0,
-  userLocation: {
-    flat: '',
-    street: '',
-    city: '',
-    country: '',
-    zip: '',
-  },
-  coords: { lat: 0, lng: 0 },
 };
 
 const orderReducer = (state = initialState, action: OrderActions): OrderState => {
   switch (action.type) {
     case OrderActionTypes.SET_PRODUCT_LIST: {
       return { ...state, productList: action.payload };
-    }
-    case OrderActionTypes.GET_USER_LOCATION: {
-      return { ...state, userLocation: action.payload };
-    }
-    case OrderActionTypes.GET_USER_COORDINATES: {
-      return { ...state, coords: action.payload };
     }
     default:
       return state;
@@ -52,18 +34,17 @@ export const setProductListAC = (payload: ProductList): SetProductListAction => 
   payload,
 });
 
-export const getUserLocationAC = (payload: UserLocation): GetUserLocationAction => ({
-  type: OrderActionTypes.GET_USER_LOCATION,
-  payload,
-});
-
-export const getUserCoordinatesAC = (payload: Coordinates): GetUserCoordinatesAction => ({
-  type: OrderActionTypes.GET_USER_COORDINATES,
-  payload,
-});
-
 export const fetchProductListAC = (): FetchProductListAction => ({
   type: OrderActionTypes.FETCH_PRODUCT_LIST,
 });
+
+function* getProductListWorker(): any {
+  const result = yield call(fetchProducts);
+  yield put(setProductListAC(result.data));
+}
+
+export function* getProductListWatcher() {
+  yield takeEvery(OrderActionTypes.FETCH_PRODUCT_LIST, getProductListWorker);
+}
 
 export default orderReducer;
