+++ src/redux/indexSaga.ts
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects';
-import { getProductListWatcher } from './ProductSaga';
-import { getUserLocationWatcher } from './LocationSaga';
+import { getProductListWatcher } from './ducks/order';
+import { getUserLocationWatcher } from './ducks/location';
 
 export function* rootwatcher() {
   yield all([getProductListWatcher(), getUserLocationWatcher()]);
