+++ src/styled/GlobalStyles.tsx
@@ -5,6 +5,5 @@ export const Global = createGlobalStyle`
     margin: 0;
     padding: 0;
     box-sizing: border-box;
-    font-family: Alef;
   }
 `;
